changed encoder to double

+++ src/edu/stuy/speed/JoeSpeed.java
@@ -9,8 +9,8 @@ package edu.stuy.speed;
  * @author admin
  */
 public interface JoeSpeed {
-    final double ENCODER_CODES_PER_REV = 360.0;
-    final double ENCODER_RPM_PER_PULSE = 60 / ENCODER_CODES_PER_REV;
+    final int ENCODER_CODES_PER_REV = 360;
+    final double ENCODER_RPM_PER_PULSE = 60.0 / ENCODER_CODES_PER_REV;
     double getRPM ();
     void setRPM (double rpm);
 }
